v1.0

+++ controller/userController.php
@@ -11,12 +11,23 @@
     }
 
     function admin() {
-      $this->view = 'admin';
-      return $this->userManagement->getUsers();
+      $result = null;
+      if ($this->isAdmin()) {
+        $this->view = 'admin';
+        $result = $this->userManagement->getUsers();
+      }
+      else {
+        $this->login();
+      }
+      return $result;
+    }
+
+    function courses() {
+      $this->view = 'courses';
     }
 
     function create() {
-      if (isset($_POST['email']) && isset($_POST['password']) && isset($_POST['name']) && isset($_POST['last_name']) && isset($_POST['phone_number']) && isset($_POST['dni']) && isset($_POST['type'])) {
+      if ($this->isAdmin() && isset($_POST['email']) && isset($_POST['password']) && isset($_POST['name']) && isset($_POST['last_name']) && isset($_POST['phone_number']) && isset($_POST['dni']) && isset($_POST['type'])) {
         $jsonData = json_decode($this->userManagement->createUser());
         if ($jsonData->success) {
           echo json_encode($jsonData);
@@ -31,7 +42,7 @@
     }
 
     function delete() {
-      if (isset($_GET['id']) && !empty($_GET['id'])) {
+      if ($this->isAdmin() && isset($_GET['id']) && !empty($_GET['id'])) {
         if ($this->userManagement->deleteUser()) {
           echo json_encode(array('success' => 1));
         }
@@ -45,7 +56,7 @@
     }
 
     function edit() {
-      if (isset($_POST['id']) && !empty($_POST['id']) && isset($_POST['email']) && isset($_POST['password']) && isset($_POST['name']) && isset($_POST['last_name']) && isset($_POST['phone_number']) && isset($_POST['dni']) && isset($_POST['type'])) {
+      if ($this->isAdmin() && isset($_POST['id']) && !empty($_POST['id']) && isset($_POST['email']) && isset($_POST['password']) && isset($_POST['name']) && isset($_POST['last_name']) && isset($_POST['phone_number']) && isset($_POST['dni']) && isset($_POST['type'])) {
         if ($this->userManagement->editUser()) {
           echo json_encode(array('success' => 1));
         }
@@ -62,6 +73,28 @@
       return $this->view;
     }
 
+    function home() {
+      $result = null;
+      if ($this->isAdmin()) {
+        $result = $this->admin();
+      }
+      else if (isset($_SESSION["type"])) {
+        $this->view = 'home';
+      }
+      else {
+        $this->login();
+      }
+      return $result;
+    }
+
+    function index() {
+      $this->view = 'index';
+    }
+
+    function isAdmin() {
+      return isset($_SESSION["type"]) && $_SESSION["type"] == 'admin';
+    }
+
     function login() {
       if (isset($_POST['email']) && isset($_POST['password'])){
         if ($this->userManagement->isUser()) {
@@ -73,6 +106,28 @@
       }
       else {
         $this->view = 'login';
+      }
+    }
+
+    function logout() {
+      if (isset($_SESSION["type"])) {
+        $_SESSION["type"] = null;
+        session_destroy();
+      }
+      $this->view = 'index';
+    }
+
+    function register() {
+      if (isset($_POST['email']) && isset($_POST['password']) && isset($_POST['type'])) {
+        if ($this->userManagement->registerUser()) {
+          echo json_encode(array('success' => 1));
+        }
+        else {
+          echo json_encode(array('success' => 0));
+        }
+      }
+      else {
+        echo json_encode(array('success' => 0));
       }
     }
   }
